+++ packages/payload-revalidate/src/types/cache/index.ts
@@ -1,3 +1,4 @@
 export type * from './walk'
+export type * from './finders'
 export type * from './tagOptions'
 export type * from './cacheOptions'